Restructured the code: - Renamed local imports to use "../" instead of "@" - Moved Main page from the components directory to the app directory - Added typescript configs
